+++ projects/ngx-uc/src/lib/ngx-uc-zoom-view/uc-zoom-view-config.ts
@@ -3,13 +3,23 @@ import {RequiredUcCssClasses, UcCssClasses} from "../uc-css-classes";
 export interface UcZoomViewConfig {
   cssClasses?: UcCssClasses;
   resetExtViewOnMouseLeave?: boolean;
+  viewPosition?: UcZoomViewPosition;
+  viewDistance?: number;
 }
 
 export interface EnforcedUcZoomViewConfig {
   cssClasses: RequiredUcCssClasses;
   resetExtViewOnMouseLeave: boolean;
+  viewPosition: UcZoomViewPosition;
+  viewDistance: number;
 }
 
+export enum UcZoomViewPosition {
+  LEFT = 'left',
+  RIGHT = 'right',
+  TOP = 'top',
+  BOTTOM = 'bottom'
+}
 
 export const UC_ZOOM_VIEW_DEFAULT_CONFIG: EnforcedUcZoomViewConfig = {
   cssClasses: {
@@ -18,5 +28,7 @@ export const UC_ZOOM_VIEW_DEFAULT_CONFIG: EnforcedUcZoomViewConfig = {
     zoomView: 'uc-img-zoom-result',
     hideLens: 'uc-hide-lens'
   },
-  resetExtViewOnMouseLeave: true
+  resetExtViewOnMouseLeave: true,
+  viewPosition: UcZoomViewPosition.RIGHT,
+  viewDistance: 0
 };